Issue 798
Extending a MySQL-side colnames filter to query signed/unsigned flag out of the DBMS in addition to the column name. This is the most efficient way to provide both meta fields via a single query to DBMS during extract time. This filter could now be renamed to something more generic like ColumnMetaFilter, but I'm leaving it as is to avoid breaking something else.